# Optimized arithmetic methods for strided triangular matrices