@jishnub 2023-12-26
Although this PR improves performance and resolves several issues, I don't really like the hacky nature of the new methods that are specifically working around bugs with using a partly initialized `A.data`. However, perhaps there is existing code that works correctly without broadcasting, so we may need to retain the general methods that forward the arithmetic operations to the parent.